# M2/trabalho web/nodejs-Cinema/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cinema Docker
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Servidor local
...
paths:
...
  /Filme/{id}/pais:
    get:
      tags:
      - Filme_Pais
      summary: Obter os países associados a um filme
      operationId: getFilmepais
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Lista de países associados ao filme
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Filme_Pais'
                x-content-type: application/json
      x-swagger-router-controller: Filmepais
    post:
      tags:
      - Filme_Pais
      summary: Associar um país a um filme
      operationId: addFilmepais
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme_Pais'
        required: true
      responses:
        "201":
          description: País associado com sucesso ao filme
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme_Pais'
      x-swagger-router-controller: Filmepais
...
components:
  schemas:
...
    Filme_Pais:
      required:
      - filme_id
      - pais
      type: object
      properties:
        filme_id:
          type: integer
          description: ID do filme
          format: int64
        pais:
          type: string
          description: País associado ao filme
      example:
        filme_id: filme_id
        pais: pais
...